requirements.txt
youtube-transcript-api
google-generativeai
python-dotenv
Flask
google-api-python-client